requirements.txt
python-dotenv==1.0.1
openai==1.59.3
streamlit==1.41.1
langchain==0.3.14
langchain-community==0.3.14
langchain-core==0.3.29
langchain-openai==0.2.14
beautifulsoup4==4.12.3
requests==2.32.3
youtube-transcript-api==0.6.3
pytube==15.0.0